Complete config.lua rework
This was a massive rework of the qb-clothing's config.lua. Anything that was commented was removed to clean up code, the comments for parts of clothes were in some other language, that was changed to english. Cadet clothing was removed from both Female/Male Police Skins, as Short Sleeves practically serve the same purpose. Doctor job was removed, instead, Doctor is now a clothing option in the Ambulance changing room. All ID's were changed so they fit standard GTAV clothes (No more aliens when changing to clothes).

Credits to Jonezkti#1990, Skymoon#1011 and myself for finding the correct clothes ID's.